+++ back-end/src/main/java/com/blog/persistence/dao/IPostDao.java
@@ -4,6 +4,6 @@ import com.blog.persistence.model.Post;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IPostDao extends JpaRepository<Post, Long> {
+public interface IPostDao extends JpaRepository<Post, Integer> {
 
 }
