+++ back-end/src/main/java/com/blog/persistence/dao/IUserDao.java
@@ -4,6 +4,6 @@ import com.blog.persistence.model.User;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IUserDao extends JpaRepository<User, Long> {
+public interface IUserDao extends JpaRepository<User, Integer> {
 
 }
